More improvements for parsing code
