+++ gramhopper/bot.py
@@ -1,11 +1,9 @@
 import logging
 from pathlib import Path
 from telegram.ext import Updater
-from gramhopper.configuration import configuration_dir
-from gramhopper.rule_handlers import rule_handlers
 from gramhopper.handlers.combined_handlers import CombinedConversationHandler
 from gramhopper.configuration import configuration_dir
-from gramhopper.configuration.parse import RuleParser #read_and_parse_rules
+from gramhopper.configuration.rules_parser import RulesParser
 
 # Enable logging
 logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.INFO)
@@ -20,7 +18,7 @@ def main():
     with open(Path(configuration_dir(), 'token.txt'), 'r') as token_file:
         bot_token = token_file.read().strip()
 
-    rule_parser = RuleParser()
+    rule_parser = RulesParser()
     rule_handlers = rule_parser.parse_file('./rules.yml')
     conversation_handler = CombinedConversationHandler(rule_handlers)
